+++ Processors/TameableProcessor.cs
@@ -68,10 +68,8 @@ sealed class TameableProcessor(ManualLogSource logger, ModConfig cfg) : Processo
                 if (!_lastMessage.TryGetValue(zdo, out var lastMessage) || (DateTimeOffset.UtcNow - lastMessage) > TimeSpan.FromSeconds(DamageText.instance.m_textDuration))
                 {
                     _lastMessage[zdo] = DateTimeOffset.UtcNow;
-                    var range = DamageText.instance.m_maxTextDistance;
-                    var playersInRange = peers.Where(x => Vector3.Distance(x.m_refPos, zdo.GetPosition()) <= range);
                     var tameness = 1f - Mathf.Clamp01(tameTimeLeft / tameTime);
-                    RPC.ShowInWorldText(playersInRange, DamageText.TextType.Normal, zdo.GetPosition(), $"$hud_tameness {tameness:P0}");
+                    RPC.ShowInWorldText(peers, DamageText.TextType.Normal, zdo.GetPosition(), $"$hud_tameness {tameness:P0}");
                 }
             }
         }